+++ extensions/transaction/transaction-atomikos/src/main/java/org/eclipse/dataspaceconnector/transaction/atomikos/AtomikosTransactionContext.java
@@ -88,5 +88,4 @@ public class AtomikosTransactionContext implements TransactionContext {
             }
         }
     }
-
 }